api-server/requirements.txt
fastapi
uvicorn
httpx

# DEV 
python-dotenv